split logic for each type of hash diff into a separate decorator
