+++ app/decorators/hash_diff_decorator/removed.rb
@@ -1,0 +1,10 @@
+module HashDiffDecorator
+  class Removed < HashDiffDecorator::Base
+    def to_html
+      C2VersionDecorator.combine_html([
+        content_tag(:code, field),
+        " was removed."
+      ])
+    end
+  end
+end
