+++ spec/decorators/hash_diff_decorator_spec.rb
@@ -1,8 +1,8 @@
 describe HashDiffDecorator do
-  describe '#to_html' do
+  describe '.html_for' do
     it "renders add events" do
-      decorated = HashDiffDecorator.new(['+', 'foo', 'bar'])
-      expect(decorated.to_html).to eq("<code>foo</code> was set to <code>bar</code>.")
+      output = HashDiffDecorator.html_for(['+', 'foo', 'bar'])
+      expect(output).to eq("<code>foo</code> was set to <code>bar</code>.")
     end
   end
 end
